+++ util/logging/messages/clienthttpresponse.go
@@ -10,78 +10,132 @@ import (
 	"github.com/opentable/sous/util/logging"
 )
 
-type clientHTTPResponse struct {
+type httpLogEntry struct {
 	logging.CallerInfo
 	logging.Level
-	method       string
-	url          string
-	server       string
-	path         string
-	parms        string
-	status       int
-	requestSize  int64
-	responseSize int64
-	dur          time.Duration
+	serverSide     bool
+	resourceFamily string
+	method         string
+	url            string
+	server         string
+	path           string
+	parms          string
+	status         int
+	requestSize    int64
+	responseSize   int64
+	dur            time.Duration
 }
 
 /*
 // ReportClientHTTPResponse reports a response recieved by Sous as a client, as provided as fields.
 func ReportClientHTTPResponseFields(logger logging.LogSink, method, server, path string, parms map[string]string, status int, dur time.Duration) {
-	m := newClientHTTPResponse(method, server, path, parms, status, dur)
+	m := newHTTPLogEntry(method, server, path, parms, status, dur)
 	logging.Deliver(m, logger)
 }
 */
 
 // ReportClientHTTPResponse reports a response recieved by Sous as a client.
-func ReportClientHTTPResponse(logger logging.LogSink, rz http.Response, dur time.Duration) {
-	url := rz.Request.URL
+func ReportClientHTTPResponse(logger logging.LogSink, rz *http.Response, resName string, dur time.Duration) {
+	m := buildHTTPLogMessage(false, rz.Request, rz.StatusCode, rz.ContentLength, resName, dur)
+	m.ExcludeMe()
+	logging.Deliver(m, logger)
+}
+
+// ReportServerHTTPResponse reports a response recieved by Sous as a client.
+// n.b. this interface subject to change
+func ReportServerHTTPResponse(logger logging.LogSink, rq *http.Request, statusCode int, contentLength int64, resName string, dur time.Duration) {
+	m := buildHTTPLogMessage(true, rq, statusCode, contentLength, resName, dur)
+	m.ExcludeMe()
+	logging.Deliver(m, logger)
+}
+
+func buildHTTPLogMessage(server bool, rq *http.Request, statusCode int, responseContentLength int64, resName string, dur time.Duration) *httpLogEntry {
+	url := rq.URL
 
 	qps := map[string]string{}
 	for k, v := range url.Query() {
 		qps[k] = strings.Join(v, ",")
 	}
 
-	m := newClientHTTPResponse(
-		rz.Request.Method,
+	m := newHTTPLogEntry(
+		server,
+		resName,
+		rq.Method,
 		url.String(),
-		rz.StatusCode,
-		rz.Request.ContentLength,
-		rz.ContentLength,
+		statusCode,
+		rq.ContentLength,
+		responseContentLength,
 		dur,
 	)
 	m.ExcludeMe()
-	logging.Deliver(m, logger)
+	return m
 }
 
-func newClientHTTPResponse(method, urlstring string, status int, rqSize, rzSize int64, dur time.Duration) *clientHTTPResponse {
+func newHTTPLogEntry(server bool, resName, method, urlstring string, status int, rqSize, rzSize int64, dur time.Duration) *httpLogEntry {
 	u, err := url.Parse(urlstring)
 	if err != nil {
 		u = &url.URL{}
 	}
 
-	return &clientHTTPResponse{
+	return &httpLogEntry{
 		Level:      logging.InformationLevel,
 		CallerInfo: logging.GetCallerInfo(logging.NotHere()),
 
-		method:       method,
-		url:          urlstring,
-		server:       u.Host,
-		path:         u.Path,
-		parms:        u.RawQuery,
-		status:       status,
-		requestSize:  rqSize,
-		responseSize: rzSize,
-		dur:          dur,
+		serverSide:     server,
+		resourceFamily: resName,
+		method:         method,
+		url:            urlstring,
+		server:         u.Host,
+		path:           u.Path,
+		parms:          u.RawQuery,
+		status:         status,
+		requestSize:    rqSize,
+		responseSize:   rzSize,
+		dur:            dur,
 	}
 }
 
-func (msg *clientHTTPResponse) MetricsTo(metrics logging.MetricsSink) {
-	metrics.UpdateTimer("http-request-duration", msg.dur)
+func (msg *httpLogEntry) MetricsTo(metrics logging.MetricsSink) {
+	side := "client"
+	if msg.serverSide {
+		side = "server"
+	}
+	host := strings.Replace(msg.server, ".", "_", -1)
+
+	nameGen := func(pat string, fields ...interface{}) []string {
+		base := fmt.Sprintf("%s-%s-%s", side, msg.method, pat)
+		if len(fields) > 0 {
+			base = fmt.Sprintf(base, fields...)
+		}
+
+		// put the metrics into 3 buckets:
+		// by "resource family" - a way of aggregating URLs
+		// by hostname (with s/./_/g for graphite's comfort)
+		// by the pair of (resource, host)
+		return []string{
+			fmt.Sprintf("%s.%s", base, msg.resourceFamily),
+			fmt.Sprintf("%s.%s", base, host),
+			fmt.Sprintf("%s.%s.%s", base, host, msg.resourceFamily),
+		}
+	}
+
+	for _, name := range nameGen("http-request-duration") {
+		metrics.UpdateTimer(name, msg.dur)
+	}
+
+	for _, name := range nameGen("http-status.%d", msg.status) {
+		metrics.IncCounter(name, 1)
+	}
+
+	for _, name := range nameGen("http-request-size") {
+		metrics.UpdateSample(name, msg.requestSize)
+	}
 }
 
-func (msg *clientHTTPResponse) EachField(f logging.FieldReportFn) {
-	f("@loglov3-otl", "http-v1")
-	f("incoming", false)
+func (msg *httpLogEntry) EachField(f logging.FieldReportFn) {
+	f("@loglov3-otl", "sous-http-v1")
+	f("resource-family", msg.resourceFamily)
+	f("incoming", msg.serverSide)
 	f("method", msg.method)
 	f("status", msg.status)
 	f("duration", int64(msg.dur/time.Microsecond))
@@ -98,6 +152,6 @@ func (msg *clientHTTPResponse) EachField(f logging.FieldReportFn) {
 	msg.CallerInfo.EachField(f)
 }
 
-func (msg *clientHTTPResponse) Message() string {
+func (msg *httpLogEntry) Message() string {
 	return fmt.Sprintf("%d", msg.status)
 }